amazonfireguard/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.9.1
requests==2.26.0
pandas==1.3.3
geopandas==0.10.2
dash==2.0.0
python-dotenv==0.19.0
scikit-learn==1.0.1
joblib==1.1.0